requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
